Move some tests to correct subfolder/file (#1312)
* move some tests to trainer file

* fix imports
